Move files into folders or delete
