--- test_driver/firebase_emulators_test.dart
@@ -1,34 +1,0 @@
-import 'package:flutter_driver/flutter_driver.dart';
-import 'package:test/test.dart';
-
-void main() {
-  group('CrowdLeague App', () {
-    // First, define the Finders and use them to locate widgets from the
-    // test suite. Note: the Strings provided to the `byValueKey` method must
-    // be the same as the Strings we used for the Keys in step 1.
-    // final counterTextFinder = find.byValueKey('counter');
-    // final buttonFinder = find.byValueKey('increment');
-
-    final buttonFinder = find.byType('AccountButton');
-
-    FlutterDriver driver;
-
-    // Connect to the Flutter driver before running any tests.
-    setUpAll(() async {
-      driver = await FlutterDriver.connect();
-    });
-
-    // Close the connection to the driver after the tests have completed.
-    tearDownAll(() async {
-      // Commented out so app can be manually used after test finishes.
-      // if (driver != null) {
-      //   await driver.close();
-      // }
-    });
-
-    test('Sign in', () async {
-      await driver.waitFor(buttonFinder);
-      await driver.tap(buttonFinder);
-    });
-  });
-}
