--- test_driver/navigation_app.dart
@@ -1,8 +1,0 @@
-import 'package:crowdleague/widgets/app/crowd_league_app.dart';
-import 'package:flutter/material.dart';
-import 'package:flutter_driver/driver_extension.dart';
-
-void main() {
-  enableFlutterDriverExtension();
-  runApp(CrowdLeagueApp());
-}
